# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /api/users/{id}:
    get:
      tags:
      - user-rest-controller
      summary: Gets a user by its id
      operationId: getUser
      parameters:
      - name: id
        in: path
        description: id of the user to be searched
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Found the user
        "404":
          description: User not found or doesn't have permission to access it
        "400":
          description: Invalid id supplied
        "401":
          description: User not authorized
    put:
      tags:
      - user-rest-controller
      summary: Update a user by its id
      operationId: putUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
        required: true
      responses:
        "200":
          description: User updated correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: User not updated
        "403":
          description: User not authorized
    delete:
      tags:
      - user-rest-controller
      summary: Delete a user by its id
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: User not deleted
        "404":
          description: User not found
        "200":
          description: User deleted correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "403":
          description: User not authorized
  /api/users/{id}/image:
    get:
      tags:
      - user-rest-controller
      summary: Gets the image of a user by its id
      operationId: getUserImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid id supplied
        "404":
          description: "User not found, user image not found or doesn't have permission\
            \ to access it"
        "401":
          description: User not authorized
        "200":
          description: Found the user image
    put:
      tags:
      - user-rest-controller
      summary: Updates the image of a user
      operationId: postUserImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "201":
          description: Image created correctly
        "401":
          description: User not authorized
        "404":
          description: User not found
        "403":
          description: User not authorized
        "400":
          description: Bad request
        "204":
          description: Image updated correctly
...
  /api/users:
    post:
      tags:
      - user-rest-controller
      summary: Registers a new user
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
        required: true
      responses:
        "409":
          description: User already exists
        "400":
          description: "Bad request, maybe one of the user attributes is missing or\
            \ the type is not valid"
        "201":
          description: User registered correctly
...
  /api/users/me:
    get:
      tags:
      - user-rest-controller
      summary: Gets the logged user
      operationId: me
      responses:
        "200":
          description: Found the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "401":
          description: User not authorized
...
components:
  schemas:
    NewUser:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        encodedPassword:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
...